feat(bastion-protocol): Added other realms
